tools/01-real-estate/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "openclaw-real-estate"
version = "1.0.0"
description = "Real Estate Dashboard — PropertyGPT, ListingSync, TenancyDoc, ViewingBot"
requires-python = ">=3.11"
dependencies = [
    "openclaw-shared",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "psutil>=5.9",
    # RAG / PropertyGPT
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    # Documents / TenancyDoc
    "python-docx>=1.1",
    "reportlab>=4.1",
    "PyPDF2>=3.0",
    "pdfplumber>=0.11",
    # Images / ListingSync
    "Pillow>=10.0",
    "pillow-heif>=0.18",
    # Browser / ListingSync
    "playwright>=1.48",
    # Calendar / ViewingBot
    "icalendar>=5.0",
    "dateparser>=1.2",
    "python-dateutil>=2.9",
    # Data
    "pandas>=2.2",
    "numpy>=1.26",
]

[project.optional-dependencies]
mlx = ["openclaw-shared[mlx]"]
messaging = ["openclaw-shared[messaging]"]
macos = ["pyobjc-framework-Vision>=10.0", "pyobjc-framework-EventKit>=10.0"]
all = ["openclaw-real-estate[mlx,messaging,macos]"]

[tool.setuptools.packages.find]
include = ["real_estate*"]
